+++ src/main/java/com/DTISE/ShelfMasterBE/infrastructure/cart/dto/GetCartResponse.java
@@ -15,7 +15,7 @@ public class GetCartResponse {
 
     private Long userId;
     private List<CartItemResponse> cartItems;
-    private Integer totalQuantity;
+    private Long totalQuantity;
     private BigDecimal totalPrice;
     private BigDecimal totalWeight;
